% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_signatures}
\alias{assign_signatures}
\title{Assign signatures based on cosine similarities}
\usage{
assign_signatures(sim_mat)
}
\arguments{
\item{sim_mat}{similarity matrix}
}
\value{
matrix
}
\description{
Assign signatures based on cosine similarities
}
