% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesNMF.R
\name{bayesNMF}
\alias{bayesNMF}
\title{Run bayesNMF}
\usage{
bayesNMF(
  data,
  rank,
  likelihood = "poisson",
  prior = "truncnormal",
  rank_method = "SBFI",
  MH = likelihood == "poisson" & prior \%in\% c("truncnormal", "exponential"),
  convergence_control = new_convergence_control(),
  prop_temp = 0.2,
  post_warmup = 2 * convergence_control$MAP_over,
  output_dir = paste0("nmf_", likelihood, "_", prior),
  overwrite = FALSE,
  hyperprior_params = list(),
  init_prior_params = list(),
  init_params = list(),
  periodic_save = TRUE,
  save_all_samples = TRUE
)
}
\arguments{
\item{data}{matrix, counts matrix (rows are variables, columns are samples)}

\item{rank}{integer, number of signatures or range of ranks to consider}

\item{likelihood}{string, likelihood distribution for M, one of "poisson", "normal" (default "poisson")}

\item{prior}{string, prior distribution for P and E, one of "truncnormal", "exponential", "gamma" (default "truncnormal")}

\item{rank_method}{string, method to determine rank, one of "SBFI", "BFI", "BIC" (heuristic) (default "SBFI")}

\item{MH}{boolean, if likelihood is Poisson, whether to use Metropolis-Hastings updates described in the paper (default TRUE if likelihood is Poisson and prior is "truncnormal" or "exponential")}

\item{convergence_control}{list, convergence control parameters (default new_convergence_control())}

\item{prop_temp}{float, proportion of iterations to use for tempering if learning rank (default 0.2)}

\item{post_warmup}{integer, number of iterations post-warmup to sample from (if MH = TRUE) (default 2 * convergence_control$MAP_over)}

\item{output_dir}{string, output directory, will be created if it does not exist}

\item{overwrite}{boolean, whether to overwrite existing output, will delete existing output directory if it exists (default FALSE)}

\item{hyperprior_params}{list, fixed hyperprior parameters}

\item{init_prior_params}{list, initial prior parameters}

\item{init_params}{list, initial parameters}

\item{periodic_save}{boolean, whether to save the sampler object periodically (default TRUE, set to FALSE to reduce I/O and computational time)}

\item{save_all_samples}{boolean, whether to save all samples (default TRUE, set to FALSE to reduce memory usage)}
}
\value{
bayesNMF_sampler object if a fixed rank is provided or rank_method is not "BIC", list of results, best rank, and sampler if "BIC"
}
\description{
Gibbs sampling for Bayesian Nonnegative Matrix Factorization with automatically learned rank and optional Metropolis-Hastings updates for computational efficiency
}
