% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{broadcast}
\alias{broadcast}
\title{Broadcast a vector across a matrix}
\usage{
broadcast(vec, across, of, with)
}
\arguments{
\item{vec}{vector to broadcast}

\item{across}{string, "rows" or "cols"}

\item{of}{matrix to broadcast across}

\item{with}{function to apply}
}
\value{
matrix
}
\description{
read as "broadcast \link{vec} across \link{across} of \link{of} with \link{with}"
e.g., broadcast(v, "rows", M, "*") =
broadcast vector v across rows of matrix M with multiplication
}
