% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_matrix_equal}
\alias{check_matrix_equal}
\title{Check if two matrices are equal}
\usage{
check_matrix_equal(mat1, mat2)
}
\arguments{
\item{mat1}{matrix}

\item{mat2}{matrix}
}
\value{
boolean
}
\description{
Returns TRUE if two matrices are both non-null and have the same dimensions and values.
}
