% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_vector_equal}
\alias{check_vector_equal}
\title{Check if two vectors are equal}
\usage{
check_vector_equal(vec1, vec2)
}
\arguments{
\item{vec1}{vector}

\item{vec2}{vector}
}
\value{
boolean
}
\description{
Returns TRUE if two vectors are both non-null and have the same length and values.
}
