% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trash.R
\name{compare_CIs}
\alias{compare_CIs}
\title{Compare credible intervals of two samplers with paired t-test between CI widths
separately for P and E matrices}
\usage{
compare_CIs(sampler1, sampler2, name1, name2)
}
\arguments{
\item{sampler1}{bayesNMF_sampler object}

\item{sampler2}{bayesNMF_sampler object}

\item{name1}{string, name of the first sampler}

\item{name2}{string, name of the second sampler}
}
\value{
data.frame with correlation and paired t-test statistics and p-values
}
\description{
Compare credible intervals of two samplers with paired t-test between CI widths
separately for P and E matrices
}
