% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{download_cosmic}
\alias{download_cosmic}
\title{Download COSMIC reference SBS signatures matrix v3.3.1 from the web}
\usage{
download_cosmic()
}
\value{
matrix
}
\description{
Download COSMIC reference SBS signatures matrix v3.3.1 from the web
}
\examples{
cosmic <- download_cosmic()
}
