% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_cosmic}
\alias{get_cosmic}
\title{Get COSMIC reference signatures matrix}
\usage{
get_cosmic()
}
\value{
matrix
}
\description{
Get COSMIC reference signatures matrix
}
\examples{
cosmic <- get_cosmic()
}
