% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_cosmic_colors}
\alias{get_cosmic_colors}
\title{Get a named list of colors used for COSMIC signature plots}
\usage{
get_cosmic_colors()
}
\value{
list
}
\description{
Get a named list of colors used for COSMIC signature plots
}
\examples{
COSMIC_colors <- get_cosmic_colors()
}
