% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_truncnormal_single.R
\name{get_exponential_prior_parameters}
\alias{get_exponential_prior_parameters}
\title{Title}
\usage{
get_exponential_prior_parameters(
  dims,
  lambda_p = sqrt(dims$N/100),
  Lambda_p = matrix(lambda_p, nrow = dims$K, ncol = dims$N),
  lambda_e = sqrt(N/100),
  Lambda_e = matrix(lambda_e, nrow = dims$N, ncol = dims$G),
  alpha = 0.1,
  Alpha = rep(alpha, dims$K),
  beta = 2,
  Beta = rep(beta, dims$K)
)
}
\arguments{
\item{lambda_p}{see `Lambda_p`}

\item{Lambda_p}{rate parameter for the exponential prior on `P`, matrix
size K x N. Defaults to all same value `lambda_p`}

\item{lambda_e}{see `Lambda_e`}

\item{Lambda_e}{rate parameter for the exponential prior on `E`, matrix
size N x G. Defaults to all same value `lambda_e`}

\item{alpha}{see `Alpha`}

\item{Alpha}{shape parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `alpha`}

\item{beta}{see `Beta`}

\item{Beta}{rate parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `beta`}
}
\description{
Title
}
