% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_heatmap}
\alias{get_heatmap}
\title{Plot a heatmap of cosine similarities between two matrices}
\usage{
get_heatmap(est_P, true_P, which = "cols")
}
\arguments{
\item{est_P}{estimated P (signatures matrix)}

\item{true_P}{true P (signatures matrix)}

\item{which}{string, one of c("rows","cols")}
}
\value{
ggplot object
}
\description{
Plot a heatmap of cosine similarities between two matrices
}
