% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_KLDiv}
\alias{get_KLDiv}
\title{Compute KL Divergence}
\usage{
get_KLDiv(M, Mhat)
}
\arguments{
\item{M}{matrix, K x G}

\item{Mhat}{reconstructed matrix, K x G}
}
\value{
scalar
}
\description{
Compute KL Divergence
}
