% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_truncnormal_single.R
\name{get_loglik_normal}
\alias{get_loglik_normal}
\title{get log likelihood}
\usage{
get_loglik_normal(M, Theta, dims)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}

\item{dims}{list of dimensions}
}
\value{
scalar
}
\description{
get log likelihood
}
