% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_gamma_single.R
\name{get_loglik_poisson_gamma_single}
\alias{get_loglik_poisson_gamma_single}
\title{get log likelihood}
\usage{
get_loglik_poisson_gamma_single(M, Theta, dims, logfac)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}

\item{dims}{list of dimensions}
}
\value{
scalar
}
\description{
get log likelihood
}
