% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mean}
\alias{get_mean}
\title{Get element-wise mean of a list of matrices}
\usage{
get_mean(matrix_list)
}
\arguments{
\item{matrix_list}{list of matrices}
}
\value{
matrix
}
\description{
Get element-wise mean of a list of matrices
}
