% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mode}
\alias{get_mode}
\title{Get mode of a list of matrices}
\usage{
get_mode(matrix_list)
}
\arguments{
\item{matrix_list}{list of matrices}
}
\value{
named list with mode ('matrix') and indices ('idx')
}
\description{
Get mode of a list of matrices
}
