% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_exponential_single.R
\name{get_mu_sigmasq_En}
\alias{get_mu_sigmasq_En}
\title{Get mu and sigmasq for E[n,]}
\usage{
get_mu_sigmasq_En(n, M, Theta)
}
\arguments{
\item{n}{signature index}

\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}
}
\value{
list of two items, mu and sigmasq
}
\description{
Get mu and sigmasq for E[n,]
}
