% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_truncnormal_single.R
\name{get_mu_sigmasq_Pn_normal_truncnormal}
\alias{get_mu_sigmasq_Pn_normal_truncnormal}
\title{get mu and sigmasq for P[,n]}
\usage{
get_mu_sigmasq_Pn_normal_truncnormal(n, M, Theta)
}
\arguments{
\item{n}{signature index}

\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}
}
\value{
list of two items, mu and sigmasq
}
\description{
get mu and sigmasq for P[,n]
}
