% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_exponential_single.R
\name{get_RMSE}
\alias{get_RMSE}
\title{get RMSE}
\usage{
get_RMSE(M, Theta)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}
}
\value{
scalar
}
\description{
get RMSE
}
