% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sim_mat}
\alias{get_sim_mat}
\title{Get similarity matrix}
\usage{
get_sim_mat(est_P, true_P)
}
\arguments{
\item{est_P}{estimated P (signatures matrix)}

\item{true_P}{true P (signatures matrix)}
}
\value{
matrix
}
\description{
Get similarity matrix
}
