% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_truncnormal_single.R
\name{get_truncnorm_prior_parameters}
\alias{get_truncnorm_prior_parameters}
\title{Title}
\usage{
get_truncnorm_prior_parameters(
  dims,
  mu_p = sqrt(100/dims$N),
  Mu_p = matrix(mu_p, nrow = dims$K, ncol = dims$N),
  sigmasq_p = mu_p,
  Sigmasq_p = matrix(sigmasq_p, nrow = dims$K, ncol = dims$N),
  mu_e = sqrt(100/dims$N),
  Mu_e = matrix(mu_e, nrow = dims$N, ncol = dims$G),
  sigmasq_e = mu_e,
  Sigmasq_e = matrix(sigmasq_e, nrow = dims$N, ncol = dims$G),
  alpha = 0.1,
  Alpha = rep(alpha, dims$K),
  beta = 2,
  Beta = rep(beta, dims$K),
  a = 0.8,
  b = 0.8
)
}
\arguments{
\item{mu_p}{see `Mu_p`}

\item{Mu_p}{mean for the truncated normal prior on `P`, matrix
size K x N. Defaults to all same value `mu_p`}

\item{sigmasq_p}{see `Sigmasq_p`}

\item{Sigmasq_p}{variance for the truncated normal prior on `P`, matrix
size K x N. Defaults to all same value `sigmasq_p`}

\item{mu_e}{see `Mu_e`}

\item{Mu_e}{mean for the truncated normal prior on `E`, matrix
size N x G. Defaults to all same value `mu_e`}

\item{sigmasq_e}{see `Sigmasq_e`}

\item{Sigmasq_e}{variance for the truncated normal prior on `P`, matrix
size N x G. Defaults to all same value `sigmasq_e`}

\item{alpha}{see `Alpha`}

\item{Alpha}{shape parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `alpha`}

\item{beta}{see `Beta`}

\item{Beta}{rate parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `beta`}
}
\description{
Title
}
