% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hungarian_assignment}
\alias{hungarian_assignment}
\title{Assign two sets of factors based on cosine similarities}
\usage{
hungarian_assignment(
  estimated_P,
  reference_P = "cosmic",
  which = "cols",
  keep_all_est = TRUE,
  keep_all_ref = FALSE,
  return_mat = FALSE
)
}
\arguments{
\item{estimated_P}{matrix, estimated factors}

\item{reference_P}{matrix, reference factors or "cosmic" to use COSMIC reference SBS signatures v3.3.1}

\item{which}{string, which dimension to compute similarity over, one of c("rows","cols") (default "cols")}

\item{keep_all_est}{boolean, whether to keep all estimated factors (default TRUE)}

\item{keep_all_ref}{boolean, whether to keep all reference factors (default FALSE)}

\item{return_mat}{boolean, whether to return the similarity matrix (default FALSE)}
}
\value{
aligned similarity matrix with rows as estimated signatures and columns as reference signatures if return_mat is TRUE, otherwise data frame with columns:
\itemize{
\item sig_est: estimated signature index or name
\item sig_ref: assigned reference signature index or name
\item cos_sim: cosine similarity between the estimated and assigned reference signatures
}
}
\description{
Relies on the Hungarian algorithm (RcppHungarian::HungarianSolver) to maximize total cosine similarity between estimated and assigned reference signatures.
}
