% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_table}
\alias{log_table}
\title{Convert a table to a string for logging}
\usage{
log_table(tab)
}
\arguments{
\item{tab}{table to convert}
}
\value{
string
}
\description{
Convert a table to a string for logging
}
