% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{new_convergence_control}
\alias{new_convergence_control}
\title{Create new convergence control object}
\usage{
new_convergence_control(
  MAP_over = 1000,
  MAP_every = 100,
  tol = 0.001,
  Ninarow_nochange = 10,
  Ninarow_nobest = 20,
  miniters = 1000,
  maxiters = 10000
)
}
\arguments{
\item{MAP_over}{integer, number of samples to average over for MAP}

\item{MAP_every}{integer, how often (in samples) to compute MAP for evaluation}

\item{tol}{numeric, tolerance for what is a meaninful "percent change"
in MAP metric}

\item{Ninarow_nobest}{integer, convergence may be determined by the number
of MAPs in a row with no new "best"}

\item{maxiters}{integer, absolute maximum number of samples to explore}

\item{Ninarow_change}{integer, convergence may be determined by the number
of MAPs in a row with no change}

\item{metric}{string, one of c('loglikelihood','RMSE','KL')}
}
\value{
list
}
\description{
Create new convergence control object
}
