% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_truncnormal_single.R
\name{nmf_normal}
\alias{nmf_normal}
\title{Run gibbs sampler for single-study NMF with normal likelihood}
\usage{
nmf_normal(
  M,
  N = NULL,
  max_N = NULL,
  A = NULL,
  P = NULL,
  E = NULL,
  sigmasq = NULL,
  prior = "truncnormal",
  prior_parameters = NULL,
  niters = 2000,
  burn_in = round(2 * niters/3),
  logevery = 100,
  file = paste0("nmf_normal_", prior),
  overwrite = FALSE,
  true_P = NULL
)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{N}{number of signatures}

\item{P}{(optional) initial signatures matrix, K x N}

\item{E}{(optional) initial exposure matrix, N x G}

\item{sigmasq}{(optinal) initial variance vector, length K}

\item{prior}{string, one of c('truncnormal','exponential')}

\item{niters}{how many iterations the Gibbs sampler is run}

\item{burn_in}{the first `burn_in` iterations will be ignored when computing
MAP estimate}

\item{logevery}{the log, save, and plot files will be updated every
`logevery` iterations}

\item{file}{file name without extension of log, save, and plot files}

\item{overwrite}{if `overwrite = TRUE`, the log, safe, and plot files of
previous runs with the same `file` will be overwritten}

\item{true_P}{(optional) true signatures matrix P to compare to in a heatmap}
}
\value{
list
}
\description{
Run gibbs sampler for single-study NMF with normal likelihood
}
