% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_exponential_single.R
\name{nmf_normal_exponential}
\alias{nmf_normal_exponential}
\title{Run gibbs sampler for normal-exponential single-study NMF}
\usage{
nmf_normal_exponential(
  M,
  N,
  P = NULL,
  E = NULL,
  sigmasq = NULL,
  niters = 1000,
  burn_in = 500,
  logevery = 100,
  file = "normal_exponential_gibbs_sampler",
  overwrite = FALSE,
  lambda_p = 5,
  Lambda_p = matrix(lambda_p, nrow = dim(M)[1], ncol = N),
  lambda_e = 3,
  Lambda_e = matrix(lambda_e, nrow = N, ncol = dim(M)[2]),
  alpha = 0.1,
  Alpha = rep(alpha, dim(M)[1]),
  beta = 2,
  Beta = rep(beta, dim(M)[1]),
  true_P = NULL
)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{N}{number of signatures}

\item{P}{(optional) initial signatures matrix, K x N}

\item{E}{(optional) initial exposure matrix, N x G}

\item{sigmasq}{(optinal) initial variance vector, length K}

\item{niters}{how many iterations the Gibbs sampler is run}

\item{burn_in}{the first `burn_in` iterations will be ignored when computing
MAP estimate}

\item{logevery}{the log, save, and plot files will be updated every
`logevery` iterations}

\item{file}{file name without extension of log, save, and plot files}

\item{overwrite}{if `overwrite = TRUE`, the log, safe, and plot files of
previous runs with the same `file` will be overwritten}

\item{lambda_p}{see `Lambda_p`}

\item{Lambda_p}{rate parameter for the exponential prior on `P`, matrix
size K x N. Defaults to all same value `lambda_p`}

\item{lambda_e}{see `Lambda_e`}

\item{Lambda_e}{rate parameter for the exponential prior on `E`, matrix
size N x G. Defaults to all same value `lambda_e`}

\item{alpha}{see `Alpha`}

\item{Alpha}{shape parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `alpha`}

\item{beta}{see `Beta`}

\item{Beta}{rate parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `beta`}

\item{true_P}{(optional) true signatures matrix P to compare to in a heatmap}
}
\value{
list
}
\description{
Run gibbs sampler for normal-exponential single-study NMF
}
