% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_truncnormal_single.R
\name{nmf_normal_truncnormal}
\alias{nmf_normal_truncnormal}
\title{Run gibbs sampler for normal-exponential single-study NMF}
\usage{
nmf_normal_truncnormal(
  M,
  N,
  P = NULL,
  E = NULL,
  sigmasq = NULL,
  niters = 1000,
  burn_in = 500,
  logevery = 100,
  file = "nmf_normal_exponential",
  overwrite = FALSE,
  mu_p = 1/5,
  Mu_p = matrix(mu_p, nrow = dim(M)[1], ncol = N),
  sigmasq_p = 1/25,
  Sigmasq_p = matrix(sigmasq_p, nrow = dim(M)[1], ncol = N),
  mu_e = 1/3,
  Mu_e = matrix(mu_e, nrow = N, ncol = dim(M)[2]),
  sigmasq_e = 1/9,
  Sigmasq_e = matrix(sigmasq_e, nrow = N, ncol = dim(M)[2]),
  alpha = 0.1,
  Alpha = rep(alpha, dim(M)[1]),
  beta = 2,
  Beta = rep(beta, dim(M)[1]),
  true_P = NULL
)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{N}{number of signatures}

\item{P}{(optional) initial signatures matrix, K x N}

\item{E}{(optional) initial exposure matrix, N x G}

\item{sigmasq}{(optinal) initial variance vector, length K}

\item{niters}{how many iterations the Gibbs sampler is run}

\item{burn_in}{the first `burn_in` iterations will be ignored when computing
MAP estimate}

\item{logevery}{the log, save, and plot files will be updated every
`logevery` iterations}

\item{file}{file name without extension of log, save, and plot files}

\item{overwrite}{if `overwrite = TRUE`, the log, safe, and plot files of
previous runs with the same `file` will be overwritten}

\item{mu_p}{see `Mu_p`}

\item{Mu_p}{mean for the truncated normal prior on `P`, matrix
size K x N. Defaults to all same value `mu_p`}

\item{sigmasq_p}{see `Sigmasq_p`}

\item{Sigmasq_p}{variance for the truncated normal prior on `P`, matrix
size K x N. Defaults to all same value `sigmasq_p`}

\item{mu_e}{see `Mu_e`}

\item{Mu_e}{mean for the truncated normal prior on `E`, matrix
size N x G. Defaults to all same value `mu_e`}

\item{sigmasq_e}{see `Sigmasq_e`}

\item{Sigmasq_e}{variance for the truncated normal prior on `P`, matrix
size N x G. Defaults to all same value `sigmasq_e`}

\item{alpha}{see `Alpha`}

\item{Alpha}{shape parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `alpha`}

\item{beta}{see `Beta`}

\item{Beta}{rate parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `beta`}

\item{true_P}{(optional) true signatures matrix P to compare to in a heatmap}
}
\value{
list
}
\description{
Run gibbs sampler for normal-exponential single-study NMF
}
