% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_gamma_single.R
\name{nmf_poisson_gamma}
\alias{nmf_poisson_gamma}
\title{Run gibbs sampler for poisson-gamma single-study NMF}
\usage{
nmf_poisson_gamma(
  M,
  N,
  P = NULL,
  E = NULL,
  niters = 2000,
  burn_in = round(2 * niters/3),
  logevery = 100,
  file = "nmf_poisson_gamma",
  overwrite = FALSE,
  alpha_p = 10,
  Alpha_p = matrix(alpha_p, nrow = dim(M)[1], ncol = N),
  beta_p = sqrt(N),
  Beta_p = matrix(beta_p, nrow = dim(M)[1], ncol = N),
  alpha_e = 10,
  Alpha_e = matrix(alpha_e, nrow = N, ncol = dim(M)[2]),
  beta_e = sqrt(N),
  Beta_e = matrix(beta_e, nrow = N, ncol = dim(M)[2]),
  true_P = NULL
)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{N}{number of signatures}

\item{P}{(optional) initial signatures matrix, K x N}

\item{E}{(optional) initial exposure matrix, N x G}

\item{niters}{how many iterations the Gibbs sampler is run}

\item{burn_in}{the first `burn_in` iterations will be ignored when computing
MAP estimate}

\item{logevery}{the log, save, and plot files will be updated every
`logevery` iterations}

\item{file}{file name without extension of log, save, and plot files}

\item{overwrite}{if `overwrite = TRUE`, the log, safe, and plot files of
previous runs with the same `file` will be overwritten}

\item{alpha_p}{see `Alpha_p`}

\item{Alpha_p}{shape parameter for the gamma prior on `P`, matrix
size K x N. Defaults to all same value `alpha_p`}

\item{beta_p}{see `Beta_p`}

\item{Beta_p}{rate parameter for the gamma prior on `P`, matrix
size K x N. Defaults to all same value `beta_p`}

\item{alpha_e}{see `Alpha_e`}

\item{Alpha_e}{shape parameter for the gamma prior on `E`, matrix
size N x G. Defaults to all same value `alpha_e`}

\item{beta_e}{see `Beta_e`}

\item{Beta_e}{rate parameter for the gamma prior on `E`, matrix
size N x G. Defaults to all same value `beta_e`}

\item{true_P}{(optional) true signatures matrix P to compare to in a heatmap}
}
\value{
list
}
\description{
Run gibbs sampler for poisson-gamma single-study NMF
}
