% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pairwise_sim}
\alias{pairwise_sim}
\title{Pairwise cosine similarity between rows or columns of matrices}
\usage{
pairwise_sim(mat1, mat2 = "cosmic", name1 = NULL, name2 = NULL, which = "cols")
}
\arguments{
\item{mat1}{matrix, first matrix for comparison}

\item{mat2}{matrix, second matrix for comparison or "cosmic" to use COSMIC reference SBS signatures v3.3.1}

\item{name1}{string, to name rows or cols of similarity matrix}

\item{name2}{string, to name rows or cols of similarity matrix}

\item{which}{string, which dimension to compute similarity over, one of c("rows","cols")}
}
\value{
matrix
}
\description{
Pairwise cosine similarity between rows or columns of matrices
}
