% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_visualizations.R
\name{plot.bayesNMF_sampler}
\alias{plot.bayesNMF_sampler}
\title{Plot a bayesNMF_sampler object}
\usage{
\method{plot}{bayesNMF_sampler}(
  sampler,
  sigs = FALSE,
  reference_P = "cosmic",
  combine_below = 0.8,
  verbose = TRUE,
  update_saved_object = TRUE
)
}
\arguments{
\item{sampler}{bayesNMF_sampler object}

\item{sigs}{boolean, whether to plot signatures}

\item{reference_P}{matrix or "cosmic", reference signatures to align to}

\item{combine_below}{numeric, threshold for combining signatures into "other"}

\item{verbose}{boolean, whether to print verbose output listing the saved plots and file locations}

\item{update_saved_object}{boolean, whether to update the bayesNMF_sampler object with stored plots and assignments}
}
\value{
named list of ggplot2 objects
}
\description{
Plot a bayesNMF_sampler object
}
