% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_visualizations.R
\name{plot_label_switching}
\alias{plot_label_switching}
\title{Plot label switching diagnostic}
\usage{
plot_label_switching(
  sampler,
  reference_P = "cosmic",
  idx = "all",
  combine_below = 0.8
)
}
\arguments{
\item{sampler}{bayesNMF_sampler object}

\item{reference_P}{matrix or "cosmic", reference signatures to align to}

\item{idx}{vector of indices or "all", indices of iterations to plot}

\item{combine_below}{numeric, threshold for combining signatures into "other"}
}
\value{
ggplot2 object
}
\description{
Plot for each iteration, the assigned reference and inclusion status for each estimated signature.
If label switching occurs, colors will switch between estimated signatures.
This is a visual diagnostic for label switching, not a statistical test.
Recommended reference matrices include COSMIC reference signatures (default), a different literature reference, or the final MAP estimate of P if a reference does not exist.
}
