% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trash.R
\name{plot_MAP_comparison}
\alias{plot_MAP_comparison}
\title{Plot a comparison of MAPs of two samplers}
\usage{
plot_MAP_comparison(sampler1, sampler2, name1, name2)
}
\arguments{
\item{sampler1}{bayesNMF_sampler object}

\item{sampler2}{bayesNMF_sampler object}

\item{name1}{string, name of the first sampler}

\item{name2}{string, name of the second sampler}
}
\value{
ggplot2 object
}
\description{
Plot a comparison of MAPs of two samplers
}
