% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_visualizations.R
\name{plot_sig}
\alias{plot_sig}
\title{Plot a single mutational signature.}
\usage{
plot_sig(
  sampler = NULL,
  sig = NULL,
  ref = "assigned",
  reference_P = "cosmic",
  title = "",
  cosine = TRUE
)
}
\arguments{
\item{sampler}{bayesNMF_sampler object, results list}

\item{sig}{integer, index of the estimated signature to plot}

\item{ref}{integer, column name string, or one of "assigned", "best" (default "assigned")}

\item{reference_P}{matrix or "cosmic", reference factors for comparison}

\item{title}{string}

\item{cosine}{boolean, whether to report cosine similarity between estimated
and reference. Ignored if either of sig, ref are NULL.}
}
\value{
ggplot2 object
}
\description{
This function can be used to plot a reference mutational signature, an
estimated mutational signature with posterior uncertainty as points with
error bars, or both. The default reference is COSMIC, and unless otherwise
specified, the reference signature chosen to plot is the one assigned using
our posterior ensemble assignment procedure.

One of the following must be true
\itemize{
\item sampler and sig are specified
\item reference_P is specified and ref is integer, a column name, or one of "assigned", "best"
\item ref is a vector
}

Often ref = "assigned" is the same as ref = "best", but not necessarily as the assignment
procedure assumes only one estimated signature can be assigned to a reference signature, and
vise versa.
}
