% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_signature_dist}
\alias{plot_signature_dist}
\title{Plot the distribution of mutational counts across signatures for each mutation
type across all, a subset, or one subject.}
\usage{
plot_signature_dist(
  res,
  subjects = 1:ncol(res$model$M),
  reference = "cosmic",
  title = "Distribution of Signature Allocation"
)
}
\arguments{
\item{res}{bayesNMF object, results list}

\item{subjects}{integer or vector, indices of subjects to include}

\item{reference}{matrix, "cosmic", or NULL, reference signatures to align to}

\item{title}{string, title of the produced plot}
}
\value{
ggplot2 object
}
\description{
Plot the distribution of mutational counts across signatures for each mutation
type across all, a subset, or one subject.
}
\examples{
plot_signature_dist(res, title = "Plot signature distribution across all subjects")
plot_signature_dist(res, subject = 1, title = "Plot signature distribution of subject 1")
plot_signature_dist(res, subject = 1, reference = NULL, title = "Plot signature distribution of subject 1 w/o reference")
plot_signature_dist(res, subject = c(1,4,10), title = "Plot signature distribution of subjects 1, 4, and 10")
}
