% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_visualizations.R
\name{plot_similarity_heatmap}
\alias{plot_similarity_heatmap}
\title{Plot heatmap of cosine similarities}
\usage{
plot_similarity_heatmap(
  estimated_P,
  reference_P = "cosmic",
  est_names = NULL,
  ref_names = NULL,
  which = "cols",
  keep_all_est = TRUE,
  keep_all_ref = FALSE
)
}
\arguments{
\item{estimated_P}{matrix, estimated P (factors matrix)}

\item{reference_P}{matrix or "cosmic", reference factors to align to}

\item{est_names}{names of estimated factors}

\item{ref_names}{names of reference factors}

\item{which}{string, one of c("rows","cols")}

\item{keep_all_est}{boolean, whether to keep all estimated factors (versus only assigned estimated factors)}

\item{keep_all_ref}{boolean, whether to keep all reference factors (versus only assigned reference factors)}
}
\value{
ggplot object
}
\description{
Plot a heatmap of cosine similarities between two matrices
with \code{ggplot2}. Can be similarity between rows or columns with the \code{which}
parameter.
}
