% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_visualizations.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot a summary of the sampler's results}
\usage{
plot_summary(
  sampler_list,
  reference_P = "cosmic",
  title = "",
  fontsize = 20,
  keep_all_ref = FALSE
)
}
\arguments{
\item{sampler_list}{named list of bayesNMF_sampler objects}

\item{reference_P}{matrix or "cosmic", reference signatures to align to}

\item{title}{string, title of the plot}

\item{fontsize}{integer, font size}

\item{keep_all_ref}{boolean, whether to keep all reference signatures in the plot, or only the ones assigned to at least one sampler}
}
\value{
ggplot2 object
}
\description{
Plot a summary of the sampler's results, including the median contribution of each signature to the samples
and the cosine similarity between the estimated and assigned reference signatures. Each sampler is plotted as its own
x-axis tick, labeled by sampler name and sample size (G) of its dataset.
}
