% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reassign_signatures}
\alias{reassign_signatures}
\title{Reassign signatures in a similarity matrix}
\usage{
reassign_signatures(sim_mat)
}
\arguments{
\item{sim_mat}{similarity matrix between estimated and true signatures}
}
\value{
matrix
}
\description{
Reassign signatures in a similarity matrix
}
