% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{renormalize}
\alias{renormalize}
\title{Renormalize NMF results for factors to sum to 1 while maintaining product}
\usage{
renormalize(P, E)
}
\arguments{
\item{P}{matrix, dimensions K x N, factors}

\item{E}{matrix, dimensions N x G, signatures}
}
\value{
list, P and E renormalized for factors to sum to 1
}
\description{
Renormalize NMF results for factors to sum to 1 while maintaining product
}
