% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_bayesNMF}
\alias{rescale_bayesNMF}
\title{Rescale MAP and credible intervals of P, E for columns of P to sum to 1}
\usage{
rescale_bayesNMF(res)
}
\arguments{
\item{res}{bayesNMF res object}
}
\value{
rescaled bayesNMF res object
}
\description{
Rescale MAP and credible intervals of P, E for columns of P to sum to 1
}
