% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_truncnormal_single.R
\name{sample_An_normal}
\alias{sample_An_normal}
\title{Sample An}
\usage{
sample_An_normal(n, M, Theta, dims, gamma = 1)
}
\arguments{
\item{n}{integer, signature index}

\item{M}{list of matrices}

\item{Theta}{list of parameters}

\item{dims}{list of dimension values}

\item{gamma}{double, tempering parameter}
}
\value{
integer
}
\description{
Sample An
}
