% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_truncnormal_single.R
\name{sample_En_normal_truncnormal}
\alias{sample_En_normal_truncnormal}
\title{Sample E[n,]}
\usage{
sample_En_normal_truncnormal(n, M, Theta)
}
\arguments{
\item{n}{signature index}

\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}
}
\value{
vector of length G
}
\description{
Sample E[n,]
}
