% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_gamma_single.R
\name{sample_Eng_poisson_gamma_single}
\alias{sample_Eng_poisson_gamma_single}
\title{Sample E[n,g]}
\usage{
sample_Eng_poisson_gamma_single(n, g, M, Theta)
}
\arguments{
\item{n}{signature index}

\item{g}{tumor genome index}

\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}
}
\value{
vector of length G
}
\description{
Sample E[n,g]
}
