% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_gamma_single.R
\name{sample_Pkn_poisson_gamma_single}
\alias{sample_Pkn_poisson_gamma_single}
\title{sample P[k,n]}
\usage{
sample_Pkn_poisson_gamma_single(k, n, M, Theta)
}
\arguments{
\item{k}{mutation type index}

\item{n}{signature index}

\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}
}
\value{
vector length K
}
\description{
sample P[k,n]
}
