% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_exponential_single.R
\name{sample_Pn}
\alias{sample_Pn}
\title{sample P[,n]}
\usage{
sample_Pn(n, M, Theta)
}
\arguments{
\item{n}{signature index}

\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}
}
\value{
vector length K
}
\description{
sample P[,n]
}
