% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_truncnormal_single.R
\name{sample_qn_normal_truncnormal}
\alias{sample_qn_normal_truncnormal}
\title{Sample qn}
\usage{
sample_qn_normal_truncnormal(n, Theta, gamma = 1)
}
\arguments{
\item{n}{integer, signature index}

\item{Theta}{list of parameters}

\item{gamma}{double, tempering parameter}
}
\value{
double
}
\description{
Sample qn
}
