% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_exponential_single.R
\name{sample_sigmasq}
\alias{sample_sigmasq}
\title{sample sigmasq}
\usage{
sample_sigmasq(M, Theta, dims)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}

\item{dims}{list of dimensions}
}
\value{
vector length K
}
\description{
sample sigmasq
}
