% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_gamma_single.R
\name{sample_Zkg_poisson_gamma_single}
\alias{sample_Zkg_poisson_gamma_single}
\title{sample Z[k,,g]}
\usage{
sample_Zkg_poisson_gamma_single(k, g, M, Theta, dims)
}
\arguments{
\item{k}{mutation type index}

\item{g}{tumor genome index}

\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}

\item{dims}{list of dimensions}
}
\value{
vector length K
}
\description{
sample Z[k,,g]
}
