% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesNMF.R
\name{set_gamma_prior_parameters}
\alias{set_gamma_prior_parameters}
\title{Set prior parameters for Gamma prior}
\usage{
set_gamma_prior_parameters(
  dims,
  alpha_p = 10,
  Alpha_p = matrix(alpha_p, nrow = dims$K, ncol = dims$N),
  beta_p = sqrt(dims$N),
  Beta_p = matrix(beta_p, nrow = dims$K, ncol = dims$N),
  alpha_e = 10,
  Alpha_e = matrix(alpha_e, nrow = dims$N, ncol = dims$G),
  beta_e = sqrt(dims$N),
  Beta_e = matrix(beta_e, nrow = dims$N, ncol = dims$G),
  a = 0.8,
  b = 0.8
)
}
\arguments{
\item{dims}{named list of dimensions N, K, G}

\item{alpha_p}{see `Alpha_p`}

\item{Alpha_p}{shape parameter for the gamma prior on `P`, matrix
size K x N. Defaults to all same value `alpha_p`}

\item{beta_p}{see `Beta_p`}

\item{Beta_p}{rate parameter for the gamma prior on `P`, matrix
size K x N. Defaults to all same value `beta_p`}

\item{alpha_e}{see `Alpha_e`}

\item{Alpha_e}{shape parameter for the gamma prior on `E`, matrix
size N x G. Defaults to all same value `alpha_e`}

\item{beta_e}{see `Beta_e`}

\item{Beta_e}{rate parameter for the gamma prior on `E`, matrix
size N x G. Defaults to all same value `beta_e`
#'}
}
\value{
named list of prior parameters
}
\description{
Set prior parameters for Gamma prior
}
