% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_truncnorm_prior_parameters}
\alias{set_truncnorm_prior_parameters}
\title{Set prior parameters for Truncated Normal prior}
\usage{
set_truncnorm_prior_parameters(
  Theta,
  dims,
  mean_p = sqrt(100/dims$N),
  mu_p = uniroot(function(x) {
     x + sqrt(x) * dnorm(-sqrt(x))/(pnorm(-sqrt(x)) - 1) -
    mean_p
 }, interval = c(0, 100))$root,
  Mu_p = matrix(mu_p, nrow = dims$K, ncol = dims$N),
  sigmasq_p = mu_p,
  Sigmasq_p = matrix(sigmasq_p, nrow = dims$K, ncol = dims$N),
  mean_e = sqrt(100/dims$N),
  mu_e = uniroot(function(x) {
     x + sqrt(x) * dnorm(-sqrt(x))/(pnorm(-sqrt(x)) - 1) -
    mean_e
 }, interval = c(0, 100))$root,
  Mu_e = matrix(mu_e, nrow = dims$N, ncol = dims$G),
  sigmasq_e = mu_e,
  Sigmasq_e = matrix(sigmasq_e, nrow = dims$N, ncol = dims$G),
  alpha = 6,
  Alpha = rep(alpha, dims$K),
  beta = 10,
  Beta = rep(beta, dims$K),
  a = 0.8,
  b = 0.8
)
}
\arguments{
\item{Theta}{list of parameters}

\item{dims}{named list of dimensions N, K, G}

\item{mu_p}{see `Mu_p`}

\item{Mu_p}{mean for the truncated normal prior on `P`, matrix
size K x N. Defaults to all same value `mu_p`}

\item{sigmasq_p}{see `Sigmasq_p`}

\item{Sigmasq_p}{variance for the truncated normal prior on `P`, matrix
size K x N. Defaults to all same value `sigmasq_p`}

\item{mu_e}{see `Mu_e`}

\item{Mu_e}{mean for the truncated normal prior on `E`, matrix
size N x G. Defaults to all same value `mu_e`}

\item{sigmasq_e}{see `Sigmasq_e`}

\item{Sigmasq_e}{variance for the truncated normal prior on `P`, matrix
size N x G. Defaults to all same value `sigmasq_e`}

\item{alpha}{see `Alpha`}

\item{Alpha}{shape parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `alpha`}

\item{beta}{see `Beta`}

\item{Beta}{rate parameter for the inverse-gamma prior on `sigmasq`,
length K. Defaults to all same value `beta`}
}
\value{
named list of prior parameters
}
\description{
Set prior parameters for Truncated Normal prior
}
