% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{signature_asssignment_inference}
\alias{signature_asssignment_inference}
\title{Posterior inference on signature assignment}
\usage{
signature_asssignment_inference(res, reference = "cosmic")
}
\arguments{
\item{res}{bayesNMF object}

\item{reference}{matrix, "cosmic", or NULL, reference signatures to align to}
}
\value{
list, signature assignments, MAP and credible intervals for cosine similarity
}
\description{
This function performs the Hungarian algorithm on each P matrix among the
posterior samples. Each sample gets to vote for it's assignment with  cosine
similarity as voting weights.
}
\examples{
res <- readRDS("examples/plot_example.rds")
assign <- signature_asssignment_inference(res)
assign$assignment
assign$MAP$cos_sim
assign$credible_intervals$cos_sim
}
