% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{summarize_samplers}
\alias{summarize_samplers}
\title{For a named list of results, gets signature assignments, cosine similarities,
and distribution of mutations attributed to signatures in each.}
\usage{
summarize_samplers(sampler_list, reference_P = "cosmic")
}
\arguments{
\item{sampler_list}{Named list, containing one or more bayesNMF sampler objects.
Names will become identifiers in the data frame.}

\item{reference_P}{matrix, "cosmic" , or NULL, reference signatures to align to}
}
\value{
data frame with columns:
\itemize{
\item Name: name of the sampler
\item G: number of signatures
\item N: number of samples
\item K: number of mutations
\item Signature: assigned reference signature
\item Med_Contribution: median contribution of the signature to the samples among samples with at least 1 mutation attributed to the signature
\item Prop_atleast_1: proportion of samples with at least 1 mutation attributed to the signature
\item Reference_Signature: assigned reference signature (if reference_P is provided)
\item Cosine_Similarity: cosine similarity between the estimated and assigned reference signatures (if reference_P is provided)
}
}
\description{
For a named list of results, gets signature assignments, cosine similarities,
and distribution of mutations attributed to signatures in each.
}
