% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{summary.bayesNMF_sampler}
\alias{summary.bayesNMF_sampler}
\title{Summary method for bayesNMF_sampler object}
\usage{
\method{summary}{bayesNMF_sampler}(sampler, reference_P = "cosmic")
}
\arguments{
\item{sampler}{bayesNMF_sampler object}

\item{reference_P}{matrix, "cosmic", or NULL, reference signatures to align to}
}
\value{
data frame with columns:
\itemize{
\item G: number of signatures
\item N: number of samples
\item K: number of mutations
\item Signature: assigned reference signature
\item Med_Contribution: median contribution of the signature to the samples among samples with at least 1 mutation attributed to the signature
\item Prop_atleast_1: proportion of samples with at least 1 mutation attributed to the signature
\item Reference_Signature: assigned reference signature (if reference_P is provided)
\item Cosine_Similarity: cosine similarity between the estimated and assigned reference signatures (if reference_P is provided)
}
}
\description{
Summary method for bayesNMF_sampler object
}
