% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Plot trace of metrics}
\usage{
trace_plot(sampler, MAP_means = FALSE, idx = "all", save = FALSE)
}
\arguments{
\item{sampler}{bayesNMF_sampler object}

\item{MAP_means}{boolean, whether to plot means of MAP metrics}

\item{idx}{vector of indices or "all" (default), indices of iterations to plot}

\item{save}{boolean, whether to save the plot to the sampler's output directory (default FALSE)}
}
\value{
ggplot2 object
}
\description{
Plot trace of metrics for the sampler, including log posterior, log likelihood, BIC, RMSE, KL-Divergence.
If rank is being learned, also reports rank, number of functional parameters (used in BIC), and temperature.
If Metropolis-Hastings steps are used, also reports average acceptance rate for elements of $P$ and $E$ (initally all 1 to find high posterior region, then true MH samples for inference).
Note that for log likelihood and log posterior, values from the Poisson models are not comparable to those from Normal models.
}
