% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_list}
\alias{update_list}
\title{Update a list with a new value at a given index, and remove the first element if the list is too long}
\usage{
update_list(list, new_value, index, max_length)
}
\arguments{
\item{list}{list to update}

\item{new_value}{value to update}

\item{index}{index to update}

\item{max_length}{maximum length of list}
}
\value{
list
}
\description{
Update a list with a new value at a given index, and remove the first element if the list is too long
}
