% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_metrics}
\alias{update_metrics}
\title{Update list of metrics}
\usage{
update_metrics(
  metrics,
  MAP,
  iter,
  Theta,
  M_truescale,
  M,
  likelihood,
  prior,
  dims,
  logfac,
  rescale_by,
  sigmasq_type
)
}
\arguments{
\item{metrics}{list, list of metrics}

\item{MAP}{list, maximum a-posteriori estimates}

\item{iter}{integer, current iteration}

\item{Theta}{list, current state of Theta}

\item{M_truescale}{matrix, true M matrix}

\item{M}{matrix, rescaled M matrix}

\item{likelihood}{string, one of c("normal", "poisson")}

\item{prior}{string, one of c("truncnormal","exponential","gamma")}

\item{dims}{list, named list of dimensions}

\item{logfac}{vector, logfac[i] = log(i!), only needed for `likelihood = 'poisson'`}

\item{rescale_by}{double, rescale factor}

\item{sigmasq_type}{string, one of c('eq_mu','invgamma','noninformative')}
}
\value{
list, updated metrics
}
\description{
Update list of metrics
}
