% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesNMF.R
\name{bayesNMF}
\alias{bayesNMF}
\title{Bayesian Non-Negative Matrix Factorization}
\usage{
bayesNMF(
  M,
  N = NULL,
  max_N = NULL,
  likelihood = "normal",
  prior = "truncnormal",
  inits = NULL,
  fixed = NULL,
  prior_parameters = NULL,
  recovery = FALSE,
  recovery_priors = "cosmic",
  file = paste0("nmf_", likelihood, "_", prior),
  true_P = NULL,
  convergence_control = new_convergence_control(),
  store_logs = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{N}{fixed number of latent factors}

\item{max_N}{maximum number of latent factors if learning rank}

\item{likelihood}{string, one of c('normal','poisson')}

\item{prior}{string, one of c('truncnormal','exponential')}

\item{inits}{(optional) list of initial values for P and E as well as sigmasq
if `likelihood = "normal"`}

\item{fixed}{(ptional) list of parameters to fix and not include in Gibbs
updates.}

\item{prior_parameters}{list, optional specification of prior parameters}

\item{file}{file name without extension of log, save, and plot files}

\item{true_P}{(optional) true latent factors matrix P to compare to in a heatmap}

\item{convergence_control}{list, specification of convergence parameters.
See documentation for `new_convergence_control`.}

\item{store_logs}{boolean, whether to store each sample in resulting .rds file}

\item{overwrite}{if `overwrite = TRUE`, the log, safe, and plot files of
previous runs with the same `file` will be overwritten}
}
\value{
list
}
\description{
Perform single-study Bayesian NMF with the provided likelihood and prior
combination. Exact rank `N` or maximum rank `max_N` must be provided.
}
