% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_cosmic}
\alias{get_cosmic}
\title{Get COSMIC reference SBS signatures matrix v3.3.1}
\usage{
get_cosmic()
}
\value{
matrix
}
\description{
Get COSMIC reference SBS signatures matrix v3.3.1
}
\examples{
cosmic <- get_cosmic()
}
