% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_heatmap}
\alias{get_heatmap}
\title{Plot heatmap of cosine similarities}
\usage{
get_heatmap(
  est_matrix,
  ref_matrix = "cosmic",
  est_names = NULL,
  ref_names = NULL,
  which = "cols",
  keep_all = FALSE
)
}
\arguments{
\item{est_matrix}{estimated P (signatures matrix)}

\item{ref_matrix}{true P (signatures matrix)}

\item{est_names}{names of estimated signatures}

\item{ref_names}{names of true signatures}

\item{which}{string, one of c("rows","cols")}
}
\value{
ggplot object
}
\description{
Plot a heatmap of cosine similarities between two matrices
with `ggplot2`. Can be similarity between rows or columns with the `which`
parameter.
}
