% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_heatmap}
\alias{get_heatmap}
\title{Get heatmap}
\usage{
get_heatmap(est_P, true_P)
}
\arguments{
\item{est_P}{estimated P (signatures matrix)}

\item{true_P}{true P (signatures matrix)}
}
\value{
ggplot object
}
\description{
Get heatmap
}
