% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_KLDiv}
\alias{get_KLDiv}
\title{Get KL Divergence}
\usage{
get_KLDiv(M, Theta)
}
\arguments{
\item{M}{mutational catalog matrix, K x G}

\item{Theta}{list of parameters}
}
\value{
scalar
}
\description{
Get KL Divergence
}
