% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mode}
\alias{get_mode}
\title{Get the mode of a list of matrices}
\usage{
get_mode(matrix_list)
}
\arguments{
\item{matrix_list}{list of matrices}
}
\value{
matrix
}
\description{
Find the mode of a list of matrices
by converting to strings and finding the most frequent string.
This is a simple way to find the most common matrix in a list,
for example the MAP estimate of a binary matrix.
}
