% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery_priors.R
\name{get_recovery_priors}
\alias{get_recovery_priors}
\title{Get Recovery Prior Parameters}
\usage{
get_recovery_priors(
  P,
  likelihood,
  prior,
  verbose = FALSE,
  file = file.path(paste0("recovery_priors_", likelihood, "_", prior, ".rds"))
)
}
\arguments{
\item{P}{matrix, K x N}

\item{likelihood}{string, one of c('normal','poisson')}

\item{prior}{string, one of c('truncnormal','exponential','gamma')}

\item{verbose}{boolean}

\item{file}{string, path to save recovery priors to}
}
\value{
list, prior parameters corresponding to given signatures matrix
}
\description{
Builds recovery priors for a given signatures matrix `P`.
For each signature in `P`, data is simulated from that P alone and
`bayesNMF` is run with `N = 1`. Recovery priors are centered around the
estimated maximum a-posteriori learned by `bayesNMF`.
}
