% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{new_convergence_control}
\alias{new_convergence_control}
\title{Create new convergence control object}
\usage{
new_convergence_control(
  MAP_over = 1000,
  MAP_every = 100,
  tol = 0.001,
  Ninarow_nochange = 5,
  Ninarow_nobest = 10,
  miniters = 1000,
  maxiters = 5000,
  minA = 0,
  metric = "logposterior"
)
}
\arguments{
\item{MAP_over}{integer, number of samples to average over for MAP}

\item{MAP_every}{integer, how often (in samples) to compute MAP for evaluation}

\item{tol}{numeric, tolerance for what is a meaningful "change" in MAP metric (default 0.001 means 0.1\% change)}

\item{Ninarow_nochange}{integer, number of consecutive MAPs with no change before declaring convergence (default 5)}

\item{Ninarow_nobest}{integer, number of consecutive MAPs with no new "best" before declaring convergence (default 10)}

\item{miniters}{integer, minimum number of samples to consider for convergence (default 1000)}

\item{maxiters}{integer, absolute maximum number of samples to explore (default 5000)}

\item{minA}{integer, minimum number of counts A matrix must have to consider it a valid MAP (default 0)}

\item{metric}{string, one of c('loglikelihood','logposterior','RMSE','KL') (default "logposterior")}
}
\value{
list of convergence control parameters
}
\description{
Specify convergence criteria for the bayesNMF Gibbs sampler.
}
