% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{new_convergence_control}
\alias{new_convergence_control}
\title{Create new convergence control object}
\usage{
new_convergence_control(
  MAP_over = 500,
  MAP_every = 100,
  tol = 0.001,
  Ninarow_nochange = 10,
  Ninarow_nobest = 20,
  miniters = 500,
  maxiters = 2000,
  minA = 0,
  metric = "logposterior"
)
}
\arguments{
\item{MAP_over}{integer, number of samples to average over for MAP}

\item{MAP_every}{integer, how often (in samples) to compute MAP for evaluation}

\item{tol}{numeric, tolerance for what is a meaninful "percent change"
in MAP metric}

\item{Ninarow_nobest}{integer, convergence may be determined by the number
of MAPs in a row with no new "best"}

\item{maxiters}{integer, absolute maximum number of samples to explore}

\item{minA}{integer, minimum number of counts A matrix must have to consider it a valid MAP}

\item{metric}{string, one of c('loglikelihood','logposterior','RMSE','KL')}

\item{Ninarow_change}{integer, convergence may be determined by the number
of MAPs in a row with no change}
}
\value{
list
}
\description{
Specify convergence criteria for the bayesNMF Gibbs sampler.
}
