% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pairwise_sim}
\alias{pairwise_sim}
\title{Pairwise cosine similarity between rows or columns of matrices}
\usage{
pairwise_sim(mat1, mat2, name1 = "", name2 = "", which = "rows")
}
\arguments{
\item{mat1}{matrix, first matrix for comparison}

\item{mat2}{matrix, second matrix for comparison}

\item{name1}{string, to name rows of similarity matrix}

\item{name2}{string, to name columns of similarity matrix}

\item{which}{string, one of c("rows","cols")}
}
\value{
matrix
}
\description{
Pairwise cosine similarity between rows or columns of matrices
}
