% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_results}
\alias{plot_results}
\title{This function displays which reference signatures are recovered in one or
more bayesNMF results objects. It reports MAP (posterior average) cosine
similarity to the assigned reference as size, and median (across samples with
contribution > 0) MAP contributions as color. The function returns the plot
as well as a data frame with all of the visualized information.}
\usage{
plot_results(res_list, ref_matrix = "cosmic", title = "", return_df = TRUE)
}
\arguments{
\item{res_list}{Named list, containing one or more bayesNMF objects. Names will become identifiers along the top of the plot.}

\item{ref_matrix}{matrix, "cosmic", or NULL, reference signatures to align to}

\item{title}{string, title of the produced plot}

\item{return_df}{boolean, whether to return summary data frame}
}
\value{
ggplot object, or list holding ggplot object and data frame if return_df = TRUE
}
\description{
This function displays which reference signatures are recovered in one or
more bayesNMF results objects. It reports MAP (posterior average) cosine
similarity to the assigned reference as size, and median (across samples with
contribution > 0) MAP contributions as color. The function returns the plot
as well as a data frame with all of the visualized information.
}
\examples{
res <- readRDS("examples/plot_example.rds")
data <- readRDS("examples/3_64_1_cosmic.rds")
plot_results(list("Example" = res), title = "Results of a single run")
plot_results(list("Example" = res), ref_matrix = data$P, title = "Results of a single run with custom reference")
}
