% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sig}
\alias{plot_sig}
\title{Plot a single mutational signature.}
\usage{
plot_sig(
  res = NULL,
  sig = NULL,
  ref = "best",
  ref_matrix = "cosmic",
  title = "",
  cosine = TRUE
)
}
\arguments{
\item{res}{bayesNMF object, results list}

\item{sig}{integer, index of the estimated signature to plot}

\item{ref}{integer, column name string, or "best"}

\item{ref_matrix}{matrix or "cosmic", reference signatures for comparison}

\item{title}{string}

\item{cosine}{boolean, whether to report cosine similarity between estimated
and reference. Ignored if either of sig, ref are NULL.}
}
\value{
ggplot2 object
}
\description{
This function can be used to plot a reference mutational signature, an
estimated mutational signature with posterior uncertainty as points with
error bars, or both. The default reference is COSMIC, and unless otherwise
specified, the reference signature chosen to plot is that with highest
cosine similarity to the estimated signature.
}
\details{
One of the following must be true
- res and sig are specified
- ref_matrix is specified and ref is integer, a column name, or "best"
- ref is a vector
}
\examples{
res <- readRDS("examples/plot_example.rds")
data <- readRDS("examples/3_64_1_cosmic.rds")
plot_sig(res = res, sig = 1, title = "Plotting an estimated signature with the best matched COSMIC signature")
plot_sig(res = res, sig = 1, ref_matrix = data$P, title = "Plotting an estimated signature with the best matched signature from a custom reference")
plot_sig(ref = "SBS3", title = "Plotting a reference signature alone")
plot_sig(ref = runif(96), title = "Plotting a custom reference signature alone")
plot_sig(res = res, sig = 1, ref = "SBS3", title = "Plot an estimated signature with a specific COSMIC signature")
}
