% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing_visualizations.R
\name{plot_signature_dist}
\alias{plot_signature_dist}
\title{Plot estimated signature distribution}
\usage{
plot_signature_dist(
  sampler,
  subjects = 1:ncol(sampler$data),
  reference_P = "cosmic",
  title = "Distribution of Signature Allocation"
)
}
\arguments{
\item{sampler}{bayesNMF_sampler object}

\item{subjects}{integer or vector, indices of subjects to include (default is all subjects)}

\item{reference_P}{matrix, "cosmic" (default), or NULL, reference signatures to align to}

\item{title}{string, title of the produced plot}
}
\value{
ggplot2 object
}
\description{
Plot the distribution of mutational counts across signatures for each mutation
type across all, a subset, or one subject.
}
